refactor(gui)!: menu call app directly
Instead of sending menu-events over a channel, directly call the app:
Previously, all menu-interactions generated events on a channel.
Additionally, the menu kept its own state for some parts,
which was updated by the app when it should change.
The new architecture now passes a reference to the app to the menu-draw.
When it is interacted with, it calls the relevant functions directly
instead of sending an event on a channel.
State is also queried directly from the app.
The channel is still kept internally for async interactions
like selecting a file in the file chooser.
This new structure should allow easier menu-callbacks in the future.
It should also allow for unit-testing the `AppState`.

List of changes:
- Split `App` into `AppState` and `AppUi`.
  `AppState` contains the state and interface to interact/update.
  `AppUi` has the ui-elements (like the MenuBar).
- `MenuBar` now directly interacts with `AppState`.
  It queries the current state and calls the functions to update.
- Added functions previously only in menu events to `AppState`.
- Extract out `FileType` to own module.
- `Repository::list` returns iterator instead of `Vec`.